Extract AST navigation helpers into dedicated microcrate
