+++ crates/perl-ast-navigation/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "perl-ast-navigation"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+description = "AST traversal helpers for offset lookup and package context"
+license = "MIT OR Apache-2.0"
+repository.workspace = true
+homepage.workspace = true
+documentation = "https://docs.rs/perl-ast-navigation"
+readme = "README.md"
+keywords = ["perl", "ast", "navigation", "lsp"]
+categories = ["development-tools", "text-editors"]
+
+[lib]
+doctest = false
+
+[dependencies]
+perl-parser-core = { workspace = true }
+
+[dev-dependencies]
+perl-tdd-support = { workspace = true }
+
+[lints]
+workspace = true
