+++ crates/perl-ast-navigation/README.md
@@ -1,0 +1,3 @@
+# perl-ast-navigation
+
+Small SRP helper crate for AST node traversal utilities used by navigation and semantic features.
