+++ crates/perl-semantic-analyzer/src/analysis/declaration.rs
@@ -5,6 +5,7 @@
 
 use crate::ast::{Node, NodeKind};
 use crate::workspace_index::{SymKind, SymbolKey};
+pub use perl_ast_navigation::{current_package_at, find_node_at_offset, get_node_children};
 use rustc_hash::FxHashMap;
 use std::sync::Arc;
 
@@ -1072,158 +1073,5 @@ pub fn symbol_at_cursor(ast: &Node, offset: usize, current_pkg: &str) -> Option<
             Some(SymbolKey { pkg: pkg.into(), name: bare.into(), sigil: None, kind: SymKind::Sub })
         }
         _ => None,
-    }
-}
-
-/// Determines the current package context at the given offset.
-///
-/// Scans the AST backwards from the offset to find the most recent
-/// package declaration, providing proper context for symbol resolution
-/// in Perl's package-based namespace system.
-///
-/// # Arguments
-/// * `ast` - Root AST node to search within
-/// * `offset` - Byte offset in the source document
-///
-/// # Returns
-/// Package name as string slice, defaults to "main" if no package found
-///
-/// # Performance
-/// - Search time: O(n) worst case, O(log n) typical
-/// - Memory: Returns borrowed string slice (zero-copy)
-/// - Caching: Results suitable for per-request caching
-///
-/// # Perl Parsing Context
-/// Perl package semantics:
-/// - `package Foo;` declarations change current namespace
-/// - Scope continues until next package declaration or EOF
-/// - Default package is "main" when no explicit declaration
-/// - Package names follow Perl identifier rules (`::`-separated)
-///
-/// # Examples
-/// ```rust,ignore
-/// use perl_parser::declaration::current_package_at;
-/// use perl_parser::ast::Node;
-///
-/// let ast = Node::new_root();
-/// let pkg = current_package_at(&ast, 100);
-/// println!("Current package: {}", pkg);
-/// ```
-pub fn current_package_at(ast: &Node, offset: usize) -> &str {
-    // Find the nearest package declaration before the offset
-    fn scan<'a>(node: &'a Node, offset: usize, last: &mut Option<&'a str>) {
-        if let NodeKind::Package { name, .. } = &node.kind {
-            if node.location.start <= offset {
-                *last = Some(name.as_str());
-            }
-        }
-        for child in get_node_children(node) {
-            if child.location.start <= offset {
-                scan(child, offset, last);
-            }
-        }
-    }
-
-    let mut last_pkg: Option<&str> = None;
-    scan(ast, offset, &mut last_pkg);
-    last_pkg.unwrap_or("main")
-}
-
-/// Finds the most specific AST node containing the given byte offset.
-///
-/// Performs recursive descent through the AST to locate the deepest node
-/// that encompasses the specified position. Essential for cursor-based
-/// LSP operations like go-to-definition and hover.
-///
-/// # Arguments
-/// * `node` - AST node to search within (typically root)
-/// * `offset` - Byte offset in the source document
-///
-/// # Returns
-/// * `Some(&Node)` - Deepest node containing the offset
-/// * `None` - Offset is outside the node's range
-///
-/// # Performance
-/// - Search time: O(log n) average, O(n) worst case
-/// - Memory: Zero allocations, returns borrowed reference
-/// - Spatial locality: Optimized for sequential offset queries
-///
-/// # LSP Integration
-/// Core primitive for:
-/// - Hover information: Find node for symbol details
-/// - Go-to-definition: Identify symbol under cursor
-/// - Completion: Determine context for suggestions
-/// - Diagnostics: Map error positions to AST nodes
-///
-/// # Examples
-/// ```rust,ignore
-/// use perl_parser::declaration::find_node_at_offset;
-/// use perl_parser::ast::Node;
-///
-/// let ast = Node::new_root();
-/// if let Some(node) = find_node_at_offset(&ast, 42) {
-///     println!("Found node: {:?}", node.kind);
-/// }
-/// ```
-pub fn find_node_at_offset(node: &Node, offset: usize) -> Option<&Node> {
-    if offset < node.location.start || offset > node.location.end {
-        return None;
-    }
-
-    // Check children first for more specific match
-    let children = get_node_children(node);
-    for child in children {
-        if let Some(found) = find_node_at_offset(child, offset) {
-            return Some(found);
-        }
-    }
-
-    // If no child contains the offset, return this node
-    Some(node)
-}
-
-/// Returns direct child nodes for a given AST node.
-///
-/// Provides generic access to child nodes across different node types,
-/// essential for AST traversal algorithms and recursive analysis patterns.
-///
-/// # Arguments
-/// * `node` - AST node to extract children from
-///
-/// # Returns
-/// Vector of borrowed child node references
-///
-/// # Performance
-/// - Time complexity: O(k) where k is child count
-/// - Memory: Allocates vector for child references
-/// - Typical latency: <5μs for common node types
-///
-/// # Examples
-/// ```rust,ignore
-/// use perl_parser::declaration::get_node_children;
-/// use perl_parser::ast::Node;
-///
-/// let node = Node::new_root();
-/// let children = get_node_children(&node);
-/// println!("Node has {} children", children.len());
-/// ```
-pub fn get_node_children(node: &Node) -> Vec<&Node> {
-    match &node.kind {
-        NodeKind::Program { statements } => statements.iter().collect(),
-        NodeKind::VariableDeclaration { variable, initializer, .. } => {
-            let mut children = vec![variable.as_ref()];
-            if let Some(init) = initializer {
-                children.push(init.as_ref());
-            }
-            children
-        }
-        NodeKind::Assignment { lhs, rhs, .. } => vec![lhs.as_ref(), rhs.as_ref()],
-        NodeKind::Binary { left, right, .. } => vec![left.as_ref(), right.as_ref()],
-        NodeKind::FunctionCall { args, .. } => args.iter().collect(),
-        NodeKind::Subroutine { body, .. } => {
-            vec![body.as_ref()]
-        }
-        NodeKind::ExpressionStatement { expression } => vec![expression.as_ref()],
-        _ => vec![],
     }
 }
